feat: Generalize wallet types and add chain support
The `TronWallet` type has been refactored to `CryptoWallet` to support multiple blockchain chains. A new `ChainType` enum is introduced for this purpose.

The `createAdminWallet` function in `StoreContext` now accepts a `ChainType` parameter, allowing for the creation of wallets on different chains.

The `UserDashboard` has been updated to allow users to select the chain type when creating sub-wallets.